% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PixhawkPlanner-package.R
\docType{package}
\name{PixhawkPlanner}
\alias{PixhawkPlanner}
\title{Pixhawk UAV Mission Planner}
\description{
The package provides some mission planning functionality for dealing with Unmanned Aerial Vehicles. The focus is set on an easy to use workflow for planning autonomous obstacle avoiding surveys of rtf-UAVs to retrieve aerial or spot related data. It provides flight control files 
based on the MAVlink protocol. Additionally it contains some useful tools for digitizing and data manipulation.
}
\details{
The package provides some mission planning functionality for dealing with Unmanned Aerial Vehicles based on the Pixhawk flightcontroller family
}
\note{
It is important to keep in mind that all auxilliary external binaries  like GDAL or SAGA need to be installed properly.
correctly on your system.
}
\author{
Chris Reudenbach Hanna Meyer Marvin Ludwig
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{package}
