% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAP.R
\name{readAP}
\alias{readAP}
\title{Read mavlink taskfiles}
\usage{
readAP(taskfile, spatial = FALSE)
}
\arguments{
\item{taskfile}{filepath of the task}

\item{spatial}{boolean, return as sf, default FALSE}
}
\value{
a list with 1. a data.frame with the waypoints and 2. the mavlink header
}
\description{
Reads a UAV mavlink flight task
}
\examples{
\dontrun{

mav = readAP(system.file("extdata", "MAVLINK_waypoints.txt", package = "PixhawkPlanner"), spatial = TRUE)
mapview::mapview(mav$task)

}



}
\author{
Marvin Ludwig
}
